+++ src/RedisSlimClient/Serialization/Protocol/ByteSequenceParser.cs
@@ -1,15 +1,12 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-using RedisSlimClient.Types;
+﻿using RedisSlimClient.Types;
 using RedisSlimClient.Types.Primatives;
+using System;
+using System.Collections.Generic;
 
 namespace RedisSlimClient.Serialization
 {
-    internal class RedisByteSequenceReader : IEnumerable<RedisObjectPart>, IDisposable
+    class ByteSequenceParser
     {
-        readonly IEnumerable<IByteSequence> _byteStream;
         readonly Stack<ArrayState> _arrayState;
 
         ReadState _currentState;
@@ -17,36 +14,28 @@ namespace RedisSlimClient.Serialization
 
         ArrayState _currentArrayState;
 
-        public RedisByteSequenceReader(IEnumerable<IByteSequence> byteStream)
+        public ByteSequenceParser()
         {
-            _byteStream = byteStream;
             _currentState = ReadState.Type;
             _currentType = (ResponseType.Unknown, 0, 0);
             _arrayState = new Stack<ArrayState>();
         }
 
-        public RedisByteSequenceReader(IEnumerable<ArraySegment<byte>> byteStream) 
-            : this(byteStream.Select(b => (IByteSequence)new ArraySegmentByteSequenceAdapter(b)))
+        public IEnumerable<RedisObjectPart> ReadItem(IByteSequence segment)
         {
-        }
-
-        public IEnumerator<RedisObjectPart> GetEnumerator()
-        {
-            foreach (var segment in _byteStream)
+            if (_currentState == ReadState.Type)
             {
-                if (_currentState == ReadState.Type)
-                {
-                    _currentType = segment.ToResponseType();
+                _currentType = segment.ToResponseType();
 
-                    switch (_currentType.type)
-                    {
-                        case ResponseType.BulkStringType:
-                            _currentState = ReadState.Value;
-                            continue;
-                        case ResponseType.ArrayType:
-                            yield return OpenArray(_currentType.length);
-                            continue;
-                        case ResponseType.IntType:
+                switch (_currentType.type)
+                {
+                    case ResponseType.BulkStringType:
+                        _currentState = ReadState.Value;
+                        yield break;
+                    case ResponseType.ArrayType:
+                        yield return OpenArray(_currentType.length);
+                        yield break;
+                    case ResponseType.IntType:
                         {
                             var part = YieldObjectPart(new RedisInteger(_currentType.length));
 
@@ -55,18 +44,17 @@ namespace RedisSlimClient.Serialization
                                 yield return part;
                             }
 
-                            continue;
+                            yield break;
                         }
-                    }
                 }
+            }
 
+            {
+                var part = YieldObjectPart(GetCurrentValue(segment));
+
+                if (!part.IsEmpty)
                 {
-                    var part = YieldObjectPart(GetCurrentValue(segment));
-
-                    if (!part.IsEmpty)
-                    {
-                        yield return part;
-                    }
+                    yield return part;
                 }
             }
         }
@@ -111,7 +99,7 @@ namespace RedisSlimClient.Serialization
         RedisObjectPart YieldObjectPart(RedisObject value)
         {
             _currentState = ReadState.Type;
-            
+
             if (_currentArrayState == null || value == null)
             {
                 return new RedisObjectPart
@@ -165,10 +153,6 @@ namespace RedisSlimClient.Serialization
             throw new NotSupportedException(_currentType.type.ToString());
         }
 
-        public void Dispose()
-        {
-        }
-
         enum ReadState
         {
             Type,
@@ -186,11 +170,6 @@ namespace RedisSlimClient.Serialization
             {
                 Index++;
             }
-        }
-
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
         }
     }
 }
