+++ src/RedisSlimClient/Serialization/Protocol/RedisByteFormatter.cs
