+++ src/RedisSlimClient/Serialization/Protocol/ResponseType.cs
